refactor(login): Move login logic back to Login page
ref #113
